Correctly handle pasting/covert to blocks of linked images.
Until now when we pasted link images the link was removed and just the image was pasted. This commits updates rawHandler to not extract images from the inside of an anchor and updates the image block to correctly handle the raw transform of an anchor with an image with or without a paragraph wrapper.