Try: Make input styles consistent
Maybe fixes #4963.

This takes a stab at unifying all basic input fields to inherit and have the same styles. That includes blocks that use input fields, and input fields in the sidebar, for example the alt text on an image block.

Blocks using the PlainText component (which is intentionally unstyled by default) can apply a `regular-text` style that is inherited from upstream. This is now applied to the Shortcode block, which also gets a visual enhancement.

While I believe this PR is a definite enhancement and adds much needed consistency, it should also get many eyes as it unstyles some upstream styles in order to achieve the goal. For example upstream input fields get a glowing blue border, and a drop shadow inside. This unstyles that to achieve consistency in Gutenberg.

As such, before I spend too much time on this, we should decide:

- A) what is the ideal _basic input style_?
- B) once we have that, is it okay to not inherit styles from the rest of WordPress? Should we create an upstream trac ticket to apply Gutenberg input styles to the rest of the admin?

Depending on discussions, I will amend this PR.